Split nt.go into multiple files
Group methods by domain:
- wait
- wait-for-sync
- validate
- metrics
