+++ e2e/nomostest/nt.go
@@ -16,12 +16,10 @@ package nomostest
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"reflect"
 	"strings"
 	"sync"
 	"time"
@@ -29,20 +27,12 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"kpt.dev/configsync/e2e/nomostest/gitproviders"
-	"kpt.dev/configsync/e2e/nomostest/metrics"
 	"kpt.dev/configsync/e2e/nomostest/ntopts"
 	"kpt.dev/configsync/pkg/core"
-	"kpt.dev/configsync/pkg/reposync"
-	"kpt.dev/configsync/pkg/rootsync"
-	"kpt.dev/configsync/pkg/util/repo"
 
 	"github.com/pkg/errors"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/rest"
@@ -50,17 +40,13 @@ import (
 	testmetrics "kpt.dev/configsync/e2e/nomostest/metrics"
 	"kpt.dev/configsync/e2e/nomostest/testing"
 	"kpt.dev/configsync/pkg/api/configmanagement"
-	v1 "kpt.dev/configsync/pkg/api/configmanagement/v1"
 	"kpt.dev/configsync/pkg/api/configsync"
 	"kpt.dev/configsync/pkg/api/configsync/v1beta1"
 	"kpt.dev/configsync/pkg/importer"
 	"kpt.dev/configsync/pkg/importer/filesystem"
-	"kpt.dev/configsync/pkg/kinds"
 	ocmetrics "kpt.dev/configsync/pkg/metrics"
 	"kpt.dev/configsync/pkg/reconcilermanager"
-	"kpt.dev/configsync/pkg/testing/fake"
 	"kpt.dev/configsync/pkg/webhook/configuration"
-	"sigs.k8s.io/cli-utils/pkg/kstatus/status"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -394,80 +380,6 @@ func (nt *NT) MustMergePatch(obj client.Object, patch string, opts ...client.Pat
 	}
 }
 
-// ValidateMetrics pulls the latest metrics, updates the metrics on NT and
-// executes the parameter function.
-func (nt *NT) ValidateMetrics(syncOption MetricsSyncOption, fn func() error) error {
-	if nt.MultiRepo {
-		nt.T.Log("validating metrics...")
-		var once sync.Once
-		duration, err := Retry(nt.DefaultMetricsTimeout, func() error {
-			duration, currentMetrics := nt.GetCurrentMetrics(syncOption)
-			nt.ReconcilerMetrics = currentMetrics
-			once.Do(func() {
-				// Only log this once. Afterwards GetCurrentMetrics will return immediately.
-				nt.T.Logf("waited %v for metrics to be current", duration)
-			})
-			return fn()
-		})
-		nt.T.Logf("waited %v for metrics to be valid", duration)
-		if err != nil {
-			return fmt.Errorf("validating metrics: %v", err)
-		}
-		return nil
-	}
-	return nil
-}
-
-// Validate returns an error if the indicated object does not exist.
-//
-// Validates the object against each of the passed Predicates, returning error
-// if any Predicate fails.
-func (nt *NT) Validate(name, namespace string, o client.Object, predicates ...Predicate) error {
-	err := nt.Get(name, namespace, o)
-	if err != nil {
-		return err
-	}
-	for _, p := range predicates {
-		err = p(o)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// ValidateNotFound returns an error if the indicated object exists.
-//
-// `o` must either be:
-// 1) a struct pointer to the type of the object to search for, or
-// 2) an unstructured.Unstructured with the type information filled in.
-func (nt *NT) ValidateNotFound(name, namespace string, o client.Object) error {
-	err := nt.Get(name, namespace, o)
-	if err == nil {
-		return errors.Errorf("%T %v %s/%s found", o, o.GetObjectKind().GroupVersionKind(), namespace, name)
-	}
-	if apierrors.IsNotFound(err) {
-		return nil
-	}
-	return err
-}
-
-// ValidateNotFoundOrNoMatch returns an error if the indicated object is
-// neither NotFound nor NoMatchFound (GVK not found).
-//
-// Use this instead of ValidateNotFound when deleting a CRD or APIService at the
-// same time as a custom resource, to avoid the race between possible errors.
-func (nt *NT) ValidateNotFoundOrNoMatch(name, namespace string, o client.Object) error {
-	err := nt.Get(name, namespace, o)
-	if err == nil {
-		return errors.Errorf("%T %v %s/%s found", o, o.GetObjectKind().GroupVersionKind(), namespace, name)
-	}
-	if apierrors.IsNotFound(err) || meta.IsNoMatchError(err) {
-		return nil
-	}
-	return err
-}
-
 // DefaultRootSyncObjectCount returns the number of objects expected to be in
 // the default RootSync repo, based on how many RepoSyncs it's managing.
 func (nt *NT) DefaultRootSyncObjectCount() int {
@@ -501,103 +413,6 @@ func (nt *NT) NumRepoSyncNamespaces() int {
 	return len(rsNamespaces)
 }
 
-// ValidateMultiRepoMetrics validates all the multi-repo metrics.
-// It checks all non-error metrics are recorded with the correct tags and values.
-func (nt *NT) ValidateMultiRepoMetrics(reconcilerName string, numResources int, gvkMetrics ...testmetrics.GVKMetric) error {
-	if nt.MultiRepo {
-		// Validate metrics emitted from the reconciler-manager.
-		if err := nt.ReconcilerMetrics.ValidateReconcilerManagerMetrics(); err != nil {
-			return err
-		}
-		// Validate non-typed and non-error metrics in the given reconciler.
-		if err := nt.ReconcilerMetrics.ValidateReconcilerMetrics(reconcilerName, numResources); err != nil {
-			return err
-		}
-		// Validate metrics that have a GVK "type" TagKey.
-		for _, tm := range gvkMetrics {
-			if err := nt.ReconcilerMetrics.ValidateGVKMetrics(reconcilerName, tm); err != nil {
-				return errors.Wrapf(err, "%s %s operation", tm.GVK, tm.APIOp)
-			}
-		}
-	}
-	return nil
-}
-
-// ValidateErrorMetricsNotFound validates that no error metrics are emitted from
-// any of the reconcilers.
-func (nt *NT) ValidateErrorMetricsNotFound() error {
-	if nt.MultiRepo {
-		for name := range nt.RootRepos {
-			reconcilerName := core.RootReconcilerName(name)
-			pod := nt.GetDeploymentPod(reconcilerName, configmanagement.ControllerNamespace)
-			if err := nt.ReconcilerMetrics.ValidateErrorMetrics(pod.Name); err != nil {
-				return err
-			}
-		}
-		for nn := range nt.NonRootRepos {
-			reconcilerName := core.NsReconcilerName(nn.Namespace, nn.Name)
-			pod := nt.GetDeploymentPod(reconcilerName, configmanagement.ControllerNamespace)
-			if err := nt.ReconcilerMetrics.ValidateErrorMetrics(pod.Name); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// ValidateResourceOverrideCount validates that the `resource_override_count` metric exists
-// for the correct reconciler.
-func (nt *NT) ValidateResourceOverrideCount(reconcilerType, containerName, resourceType string, count int) error {
-	if nt.MultiRepo {
-		return nt.ReconcilerMetrics.ValidateResourceOverrideCount(reconcilerType, containerName, resourceType, count)
-	}
-	return nil
-}
-
-// ValidateResourceOverrideCountMissingTags checks that the `resource_override_count` metric misses the specific the tags.
-func (nt *NT) ValidateResourceOverrideCountMissingTags(tags []metrics.Tag) error {
-	if nt.MultiRepo {
-		return nt.ReconcilerMetrics.ValidateResourceOverrideCountMissingTags(tags)
-	}
-	return nil
-}
-
-// ValidateGitSyncDepthOverrideCount validates the `git_sync_depth_override_count` metric.
-func (nt *NT) ValidateGitSyncDepthOverrideCount(count int) error {
-	if nt.MultiRepo {
-		return nt.ReconcilerMetrics.ValidateGitSyncDepthOverrideCount(count)
-	}
-	return nil
-}
-
-// ValidateNoSSLVerifyCount checks that the `no_ssl_verify_count` metric has the correct value.
-func (nt *NT) ValidateNoSSLVerifyCount(count int) error {
-	if nt.MultiRepo {
-		return nt.ReconcilerMetrics.ValidateNoSSLVerifyCount(count)
-	}
-	return nil
-}
-
-// ValidateMetricNotFound validates that a metric does not exist.
-func (nt *NT) ValidateMetricNotFound(metricName string) error {
-	if nt.MultiRepo {
-		if _, ok := nt.ReconcilerMetrics[metricName]; ok {
-			return errors.Errorf("Found an unexpected metric: %s", metricName)
-		}
-	}
-	return nil
-}
-
-// ValidateReconcilerErrors validates that the `reconciler_error` metric exists
-// for the correct reconciler pod and the tagged component has the correct value.
-func (nt *NT) ValidateReconcilerErrors(reconcilerName string, sourceCount, syncCount int) error {
-	if nt.MultiRepo {
-		pod := nt.GetDeploymentPod(reconcilerName, configmanagement.ControllerNamespace)
-		return nt.ReconcilerMetrics.ValidateReconcilerErrors(pod.Name, sourceCount, syncCount)
-	}
-	return nil
-}
-
 // DefaultRootSha1Fn is the default function to retrieve the commit hash of the root repo.
 func DefaultRootSha1Fn(nt *NT, nn types.NamespacedName) (string, error) {
 	// Get the repository this RootSync is syncing to, and ensure it is synced to HEAD.
@@ -617,236 +432,6 @@ func DefaultRepoSha1Fn(nt *NT, nn types.NamespacedName) (string, error) {
 		return "", fmt.Errorf("checked if nonexistent repo is synced")
 	}
 	return repo.Hash(), nil
-}
-
-// WaitForRepoSyncs is a convenience method that waits for all repositories
-// to sync.
-//
-// Unless you're testing pre-CSMR functionality related to in-cluster objects,
-// you should be using this function to block on ConfigSync to sync everything.
-//
-// If you want to check the internals of specific objects (e.g. the error field
-// of a RepoSync), use nt.Validate() - possibly in a Retry.
-func (nt *NT) WaitForRepoSyncs(options ...WaitForRepoSyncsOption) {
-	nt.T.Helper()
-
-	waitForRepoSyncsOptions := newWaitForRepoSyncsOptions(nt.DefaultWaitTimeout, DefaultRootSha1Fn, DefaultRepoSha1Fn)
-	for _, option := range options {
-		option(&waitForRepoSyncsOptions)
-	}
-
-	syncTimeout := waitForRepoSyncsOptions.timeout
-
-	if nt.MultiRepo {
-		if err := ValidateMultiRepoDeployments(nt); err != nil {
-			nt.T.Fatal(err)
-		}
-		for name := range nt.RootRepos {
-			syncDir := syncDirectory(waitForRepoSyncsOptions.syncDirectoryMap, RootSyncNN(name))
-			nt.WaitForSync(kinds.RootSyncV1Beta1(), name,
-				configmanagement.ControllerNamespace, syncTimeout,
-				waitForRepoSyncsOptions.rootSha1Fn, RootSyncHasStatusSyncCommit,
-				&SyncDirPredicatePair{syncDir, RootSyncHasStatusSyncDirectory})
-		}
-
-		syncNamespaceRepos := waitForRepoSyncsOptions.syncNamespaceRepos
-		if syncNamespaceRepos {
-			for nn := range nt.NonRootRepos {
-				syncDir := syncDirectory(waitForRepoSyncsOptions.syncDirectoryMap, nn)
-				nt.WaitForSync(kinds.RepoSyncV1Beta1(), nn.Name, nn.Namespace,
-					syncTimeout, waitForRepoSyncsOptions.repoSha1Fn, RepoSyncHasStatusSyncCommit,
-					&SyncDirPredicatePair{syncDir, RepoSyncHasStatusSyncDirectory})
-			}
-		}
-	} else {
-		nt.WaitForSync(kinds.Repo(), repo.DefaultName, "", syncTimeout,
-			waitForRepoSyncsOptions.rootSha1Fn, RepoHasStatusSyncLatestToken, nil)
-	}
-}
-
-// GetCurrentMetrics fetches metrics from the otel-collector ensuring that the
-// metrics have been updated for with the most recent commit hashes.
-func (nt *NT) GetCurrentMetrics(syncOptions ...MetricsSyncOption) (time.Duration, testmetrics.ConfigSyncMetrics) {
-	nt.T.Helper()
-
-	if nt.MultiRepo {
-		var metrics testmetrics.ConfigSyncMetrics
-
-		// Metrics are buffered and sent in batches to the collector.
-		// So we may have to retry a few times until they're current.
-		took, err := Retry(nt.DefaultWaitTimeout, func() error {
-			var err error
-			metrics, err = testmetrics.ParseMetrics(nt.otelCollectorPort)
-			if err != nil {
-				// Port forward again to fix intermittent "exit status 56" errors when
-				// parsing from the port.
-				nt.PortForwardOtelCollector()
-				return err
-			}
-
-			for _, syncOption := range syncOptions {
-				err = syncOption(&metrics)
-				if err != nil {
-					return err
-				}
-			}
-
-			return nil
-		})
-		if err != nil {
-			nt.T.Fatalf("unable to get latest metrics: %v", err)
-		}
-
-		nt.DebugLogMetrics(metrics)
-
-		return took, metrics
-	}
-
-	return 0, nil
-}
-
-// DebugLogMetrics logs metrics to the debug log, if enabled
-func (nt *NT) DebugLogMetrics(metrics testmetrics.ConfigSyncMetrics) {
-	if !*e2e.Debug {
-		return
-	}
-	nt.DebugLog("Logging all received metrics...")
-	for name, ms := range metrics {
-		for _, m := range ms {
-			tagsJSONBytes, err := json.Marshal(m.TagMap())
-			if err != nil {
-				nt.T.Fatalf("unable to convert latest tags to json for metric %q: %v", name, err)
-			}
-			nt.DebugLogf("Metric received: { \"Name\": %q, \"Value\": %#v, \"Tags\": %s }", name, m.Value, string(tagsJSONBytes))
-		}
-	}
-}
-
-// WaitForSync waits for the specified object to be synced.
-//
-// o returns a new object of the type to check is synced. It can't just be a
-// struct pointer as calling .Get on the same struct pointer multiple times
-// has undefined behavior.
-//
-// name and namespace identify the specific object to check.
-//
-// timeout specifies the maximum duration allowed for the object to sync.
-//
-// sha1Func is the function that dynamically computes the expected commit sha1.
-//
-// syncSha1 is a Predicate to use to tell whether the object is synced as desired.
-//
-// syncDirPair is a pair of sync dir and the corresponding predicate that tells
-// whether it is synced to the expected directory.
-// It will skip the validation if it is not provided.
-func (nt *NT) WaitForSync(gvk schema.GroupVersionKind, name, namespace string, timeout time.Duration,
-	sha1Func Sha1Func, syncSha1 func(string) Predicate, syncDirPair *SyncDirPredicatePair) {
-	nt.T.Helper()
-	// Wait for the repository to report it is synced.
-	took, err := Retry(timeout, func() error {
-		var nn types.NamespacedName
-		if namespace == "" {
-			// If namespace is empty, that is the monorepo mode.
-			// So using the sha1 from the default RootSync, not the Repo.
-			nn = DefaultRootRepoNamespacedName
-		} else {
-			nn = types.NamespacedName{
-				Name:      name,
-				Namespace: namespace,
-			}
-		}
-		sha1, err := sha1Func(nt, nn)
-		if err != nil {
-			return errors.Wrapf(err, "failed to retrieve sha1")
-		}
-		isSynced := []Predicate{syncSha1(sha1)}
-		if syncDirPair != nil {
-			isSynced = append(isSynced, syncDirPair.Predicate(syncDirPair.Dir))
-		}
-		return nt.ValidateSyncObject(gvk, name, namespace, isSynced...)
-	})
-	if err != nil {
-		nt.T.Logf("failed after %v to wait for %s/%s %v to be synced", took, namespace, name, gvk)
-
-		nt.T.Fatal(err)
-	}
-	nt.T.Logf("took %v to wait for %s/%s %v to be synced", took, namespace, name, gvk)
-
-	// Automatically renew the Client. We don't have tests that depend on behavior
-	// when the test's client is out of date, and if ConfigSync reports that
-	// everything has synced properly then (usually) new types should be available.
-	if gvk == kinds.Repo() || gvk == kinds.RepoSyncV1Beta1() {
-		nt.RenewClient()
-	}
-}
-
-// ValidateSyncObject validates the specified object satisfies the specified
-// constraints.
-//
-// gvk specifies the GroupVersionKind of the object to retrieve and validate.
-//
-// name and namespace identify the specific object to check.
-//
-// predicates are functions that return an error if the object is invalid.
-func (nt *NT) ValidateSyncObject(gvk schema.GroupVersionKind, name, namespace string,
-	predicates ...Predicate) error {
-	rObj, err := nt.scheme.New(gvk)
-	if err != nil {
-		return fmt.Errorf("%w: got unrecognized GVK %v", ErrWrongType, gvk)
-	}
-	cObj, ok := rObj.(client.Object)
-	if !ok {
-		// This means the GVK corresponded to a type registered in the Scheme
-		// which is not a valid Kubernetes object. We expect the only way this
-		// can happen is if gvk is for a List type, like NamespaceList.
-		return errors.Wrapf(ErrWrongType, "trying to wait for List type to sync: %T", cObj)
-	}
-	return nt.Validate(name, namespace, cObj, predicates...)
-}
-
-// WaitForNamespace waits for a namespace to exist and be ready to use
-func (nt *NT) WaitForNamespace(timeout time.Duration, namespace string) {
-	nt.T.Helper()
-
-	// Wait for the repository to report it is synced.
-	took, err := Retry(timeout, func() error {
-		obj := &unstructured.Unstructured{}
-		obj.SetGroupVersionKind(schema.GroupVersionKind{
-			Group:   corev1.SchemeGroupVersion.Group,
-			Version: corev1.SchemeGroupVersion.Version,
-			Kind:    "Namespace",
-		})
-		obj.SetName(namespace)
-		if err := nt.Get(namespace, "", obj); err != nil {
-			return fmt.Errorf("namespace %q GET failed: %w", namespace, err)
-		}
-		result, err := status.Compute(obj)
-		if err != nil {
-			return fmt.Errorf("namespace %q status could not be computed: %w", namespace, err)
-		}
-		if result.Status != status.CurrentStatus {
-			return fmt.Errorf("namespace %q not reconciled: status is %q: %s", namespace, result.Status, result.Message)
-		}
-		return nil
-	})
-	if err != nil {
-		nt.T.Logf("failed after %v to wait for namespace %q to be ready", took, namespace)
-		nt.T.Fatal(err)
-	}
-	nt.T.Logf("took %v to wait for namespace %q to be ready", took, namespace)
-}
-
-// WaitForNamespaces waits for namespaces to exist and be ready to use
-func (nt *NT) WaitForNamespaces(timeout time.Duration, namespaces ...string) {
-	var wg sync.WaitGroup
-	for _, namespace := range namespaces {
-		wg.Add(1)
-		go func(t time.Duration, ns string) {
-			defer wg.Done()
-			nt.WaitForNamespace(t, ns)
-		}(timeout, namespace)
-	}
-	wg.Wait()
 }
 
 // RenewClient gets a new Client for talking to the cluster.
@@ -1180,275 +765,6 @@ func (nt *NT) ForwardToFreePort(ns, pod, port string) (int, error) {
 	return localPort, nil
 }
 
-func validateRootSyncError(statusErrs []v1beta1.ConfigSyncError, syncingCondition *v1beta1.RootSyncCondition, code string, message string, expectedErrorSourceRefs []v1beta1.ErrorSource) error {
-	if err := validateError(statusErrs, code, message); err != nil {
-		return err
-	}
-	if syncingCondition == nil {
-		return fmt.Errorf("syncingCondition is nil")
-	}
-	if syncingCondition.Status == metav1.ConditionTrue {
-		return fmt.Errorf("status.conditions['Syncing'].status is True, expected false")
-	}
-	if !reflect.DeepEqual(syncingCondition.ErrorSourceRefs, expectedErrorSourceRefs) {
-		return fmt.Errorf("status.conditions['Syncing'].errorSourceRefs is %v, expected %v", syncingCondition.ErrorSourceRefs, expectedErrorSourceRefs)
-	}
-	return nil
-}
-
-func validateRepoSyncError(statusErrs []v1beta1.ConfigSyncError, syncingCondition *v1beta1.RepoSyncCondition, code string, message string, expectedErrorSourceRefs []v1beta1.ErrorSource) error {
-	if err := validateError(statusErrs, code, message); err != nil {
-		return err
-	}
-	if syncingCondition == nil {
-		return fmt.Errorf("syncingCondition is nil")
-	}
-	if syncingCondition.Status == metav1.ConditionTrue {
-		return fmt.Errorf("status.conditions['Syncing'].status is True, expected false")
-	}
-	if !reflect.DeepEqual(syncingCondition.ErrorSourceRefs, expectedErrorSourceRefs) {
-		return fmt.Errorf("status.conditions['Syncing'].errorSourceRefs is %v, expected %v", syncingCondition.ErrorSourceRefs, expectedErrorSourceRefs)
-	}
-	return nil
-}
-
-func validateError(errs []v1beta1.ConfigSyncError, code string, message string) error {
-	if len(errs) == 0 {
-		return errors.Errorf("no errors present")
-	}
-	var codes []string
-	for _, e := range errs {
-		if e.Code == code {
-			if message == "" || strings.Contains(e.ErrorMessage, message) {
-				return nil
-			}
-		}
-		codes = append(codes, e.Code)
-	}
-	return errors.Errorf("error %s not present, got %s", code, strings.Join(codes, ", "))
-}
-
-// WaitForRootSyncSourceError waits until the given error (code and message) is present on the RootSync resource
-func (nt *NT) WaitForRootSyncSourceError(rsName, code string, message string, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("RootSync %s source error code %s", rsName, code), nt.DefaultWaitTimeout,
-		func() error {
-			rs := fake.RootSyncObjectV1Beta1(rsName)
-			if err := nt.Get(rs.GetName(), rs.GetNamespace(), rs); err != nil {
-				return err
-			}
-			syncingCondition := rootsync.GetCondition(rs.Status.Conditions, v1beta1.RootSyncSyncing)
-			return validateRootSyncError(rs.Status.Source.Errors, syncingCondition, code, message, []v1beta1.ErrorSource{v1beta1.SourceError})
-		},
-		opts...,
-	)
-}
-
-// WaitForRootSyncRenderingError waits until the given error (code and message) is present on the RootSync resource
-func (nt *NT) WaitForRootSyncRenderingError(rsName, code string, message string, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("RootSync %s rendering error code %s", rsName, code), nt.DefaultWaitTimeout,
-		func() error {
-			rs := fake.RootSyncObjectV1Beta1(rsName)
-			err := nt.Get(rs.GetName(), rs.GetNamespace(), rs)
-			if err != nil {
-				return err
-			}
-			syncingCondition := rootsync.GetCondition(rs.Status.Conditions, v1beta1.RootSyncSyncing)
-			return validateRootSyncError(rs.Status.Rendering.Errors, syncingCondition, code, message, []v1beta1.ErrorSource{v1beta1.RenderingError})
-		},
-		opts...,
-	)
-}
-
-// WaitForRootSyncSyncError waits until the given error (code and message) is present on the RootSync resource
-func (nt *NT) WaitForRootSyncSyncError(rsName, code string, message string, ignoreSyncingCondition bool, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("RootSync %s rendering error code %s", rsName, code), nt.DefaultWaitTimeout,
-		func() error {
-			rs := fake.RootSyncObjectV1Beta1(rsName)
-			err := nt.Get(rs.GetName(), rs.GetNamespace(), rs)
-			if err != nil {
-				return err
-			}
-			if ignoreSyncingCondition {
-				return validateError(rs.Status.Sync.Errors, code, message)
-			}
-			syncingCondition := rootsync.GetCondition(rs.Status.Conditions, v1beta1.RootSyncSyncing)
-			return validateRootSyncError(rs.Status.Sync.Errors, syncingCondition, code, message, []v1beta1.ErrorSource{v1beta1.SyncError})
-		},
-		opts...,
-	)
-}
-
-// WaitForRepoSyncSyncError waits until the given error (code and message) is present on the RepoSync resource
-func (nt *NT) WaitForRepoSyncSyncError(ns, rsName, code string, message string, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("RepoSync %s/%s rendering error code %s", ns, rsName, code), nt.DefaultWaitTimeout,
-		func() error {
-			rs := fake.RepoSyncObjectV1Beta1(ns, rsName)
-			err := nt.Get(rs.GetName(), rs.GetNamespace(), rs)
-			if err != nil {
-				return err
-			}
-			syncingCondition := reposync.GetCondition(rs.Status.Conditions, v1beta1.RepoSyncSyncing)
-			return validateRepoSyncError(rs.Status.Sync.Errors, syncingCondition, code, message, []v1beta1.ErrorSource{v1beta1.SyncError})
-		},
-		opts...,
-	)
-}
-
-// WaitForRepoSyncSourceError waits until the given error (code and message) is present on the RepoSync resource
-func (nt *NT) WaitForRepoSyncSourceError(ns, rsName, code, message string, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("RepoSync %s/%s source error code %s", ns, rsName, code), nt.DefaultWaitTimeout,
-		func() error {
-			nt.T.Helper()
-
-			rs := fake.RepoSyncObjectV1Beta1(ns, rsName)
-			err := nt.Get(rs.GetName(), rs.GetNamespace(), rs)
-			if err != nil {
-				return err
-			}
-			syncingCondition := reposync.GetCondition(rs.Status.Conditions, v1beta1.RepoSyncSyncing)
-			return validateRepoSyncError(rs.Status.Source.Errors, syncingCondition, code, message, []v1beta1.ErrorSource{v1beta1.SourceError})
-		},
-		opts...,
-	)
-}
-
-// WaitForRepoSourceError waits until the given error (code and message) is present on the Repo resource
-func (nt *NT) WaitForRepoSourceError(code string, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("Repo source error code %s", code), nt.DefaultWaitTimeout,
-		func() error {
-			obj := &v1.Repo{}
-			err := nt.Get(repo.DefaultName, "", obj)
-			if err != nil {
-				return err
-			}
-			errs := obj.Status.Source.Errors
-			if len(errs) == 0 {
-				return errors.Errorf("no errors present")
-			}
-			var codes []string
-			for _, e := range errs {
-				if e.Code == fmt.Sprintf("KNV%s", code) {
-					return nil
-				}
-				codes = append(codes, e.Code)
-			}
-			return errors.Errorf("error %s not present, got %s", code, strings.Join(codes, ", "))
-		},
-		opts...,
-	)
-}
-
-// WaitForRepoSourceErrorClear waits until the given error code disappears from the Repo resource
-func (nt *NT) WaitForRepoSourceErrorClear(opts ...WaitOption) {
-	Wait(nt.T, "Repo source errors cleared", nt.DefaultWaitTimeout,
-		func() error {
-			obj := &v1.Repo{}
-			err := nt.Get(repo.DefaultName, "", obj)
-			if err != nil {
-				return err
-			}
-			errs := obj.Status.Source.Errors
-			if len(errs) == 0 {
-				return nil
-			}
-			var messages []string
-			for _, e := range errs {
-				messages = append(messages, e.ErrorMessage)
-			}
-			return errors.Errorf("got errors %s", strings.Join(messages, ", "))
-		},
-		opts...,
-	)
-}
-
-// WaitForRepoImportErrorCode waits until the given error code is present on the Repo resource.
-func (nt *NT) WaitForRepoImportErrorCode(code string, opts ...WaitOption) {
-	Wait(nt.T, fmt.Sprintf("Repo import error code %s", code), nt.DefaultWaitTimeout,
-		func() error {
-			obj := &v1.Repo{}
-			err := nt.Get(repo.DefaultName, "", obj)
-			if err != nil {
-				return err
-			}
-			errs := obj.Status.Import.Errors
-			if len(errs) == 0 {
-				return errors.Errorf("no errors present")
-			}
-			var codes []string
-			for _, e := range errs {
-				if e.Code == fmt.Sprintf("KNV%s", code) {
-					return nil
-				}
-				codes = append(codes, e.Code)
-			}
-			return errors.Errorf("error %s not present, got %s", code, strings.Join(codes, ", "))
-		},
-		opts...,
-	)
-}
-
-// WaitForRootSyncStalledError waits until the given Stalled error is present on the RootSync resource.
-func (nt *NT) WaitForRootSyncStalledError(rsNamespace, rsName, reason, message string) {
-	Wait(nt.T, fmt.Sprintf("RootSync %s/%s stalled error", rsNamespace, rsName), nt.DefaultWaitTimeout,
-		func() error {
-			rs := &v1beta1.RootSync{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      rsName,
-					Namespace: rsNamespace,
-				},
-				TypeMeta: fake.ToTypeMeta(kinds.RootSyncV1Beta1()),
-			}
-			if err := nt.Get(rsName, rsNamespace, rs); err != nil {
-				return err
-			}
-			stalledCondition := rootsync.GetCondition(rs.Status.Conditions, v1beta1.RootSyncStalled)
-			if stalledCondition == nil {
-				return fmt.Errorf("stalled condition not found")
-			}
-			if stalledCondition.Status != metav1.ConditionTrue {
-				return fmt.Errorf("expected status.conditions['Stalled'].status is True, got %s", stalledCondition.Status)
-			}
-			if stalledCondition.Reason != reason {
-				return fmt.Errorf("expected status.conditions['Stalled'].reason is %s, got %s", reason, stalledCondition.Reason)
-			}
-			if !strings.Contains(stalledCondition.Message, message) {
-				return fmt.Errorf("expected status.conditions['Stalled'].message to contain %s, got %s", message, stalledCondition.Message)
-			}
-			return nil
-		})
-}
-
-// WaitForRepoSyncStalledError waits until the given Stalled error is present on the RepoSync resource.
-func (nt *NT) WaitForRepoSyncStalledError(rsNamespace, rsName, reason, message string) {
-	Wait(nt.T, fmt.Sprintf("RepoSync %s/%s stalled error", rsNamespace, rsName), nt.DefaultWaitTimeout,
-		func() error {
-			rs := &v1beta1.RepoSync{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      rsName,
-					Namespace: rsNamespace,
-				},
-				TypeMeta: fake.ToTypeMeta(kinds.RepoSyncV1Beta1()),
-			}
-			if err := nt.Get(rsName, rsNamespace, rs); err != nil {
-				return err
-			}
-			stalledCondition := reposync.GetCondition(rs.Status.Conditions, v1beta1.RepoSyncStalled)
-			if stalledCondition == nil {
-				return fmt.Errorf("stalled condition not found")
-			}
-			if stalledCondition.Status != metav1.ConditionTrue {
-				return fmt.Errorf("expected status.conditions['Stalled'].status is True, got %s", stalledCondition.Status)
-			}
-			if stalledCondition.Reason != reason {
-				return fmt.Errorf("expected status.conditions['Stalled'].reason is %s, got %s", reason, stalledCondition.Reason)
-			}
-			if !strings.Contains(stalledCondition.Message, message) {
-				return fmt.Errorf("expected status.conditions['Stalled'].message to contain %s, got %s", message, stalledCondition.Message)
-			}
-			return nil
-		})
-}
-
 // SupportV1Beta1CRDAndRBAC checks if v1beta1 CRD and RBAC resources are supported
 // in the current testing cluster.
 // v1beta1 APIs for CRD and RBAC resources are deprecated in K8s 1.22.
@@ -1503,206 +819,6 @@ func (nt *NT) GetDeploymentPod(deploymentName, namespace string) *corev1.Pod {
 	pod := pods.Items[0]
 	nt.DebugLogf("Found deployment pod: %s/%s", pod.Namespace, pod.Name)
 	return &pod
-}
-
-// WaitOption is an optional parameter for Wait
-type WaitOption func(wait *waitSpec)
-
-type waitSpec struct {
-	timeout time.Duration
-	// failOnError is the flag to control whether to fail the test or not when errors occur.
-	failOnError bool
-}
-
-// WaitTimeout provides the timeout option to Wait.
-func WaitTimeout(timeout time.Duration) WaitOption {
-	return func(wait *waitSpec) {
-		wait.timeout = timeout
-	}
-}
-
-// WaitNoFail sets failOnError to false so the Wait function only logs the error but not fails the test.
-func WaitNoFail() WaitOption {
-	return func(wait *waitSpec) {
-		wait.failOnError = false
-	}
-}
-
-// Wait provides a logged wait for condition to return nil with options for timeout.
-// It fails the test on errors.
-func Wait(t testing.NTB, opName string, timeout time.Duration, condition func() error, opts ...WaitOption) {
-	t.Helper()
-
-	wait := waitSpec{
-		timeout:     timeout,
-		failOnError: true,
-	}
-	for _, opt := range opts {
-		opt(&wait)
-	}
-
-	// Wait for the repository to report it is synced.
-	took, err := Retry(wait.timeout, condition)
-	if err != nil {
-		t.Logf("failed after %v to wait for %s", took, opName)
-		if wait.failOnError {
-			t.Fatal(err)
-		}
-	}
-	t.Logf("took %v to wait for %s", took, opName)
-}
-
-// MetricsSyncOption determines where metrics will be synced to
-type MetricsSyncOption func(csm *testmetrics.ConfigSyncMetrics) error
-
-// SyncMetricsToLatestCommit syncs metrics to the latest synced commit
-func SyncMetricsToLatestCommit(nt *NT) MetricsSyncOption {
-	return func(metrics *testmetrics.ConfigSyncMetrics) error {
-		for syncName := range nt.RootRepos {
-			reconcilerName := core.RootReconcilerName(syncName)
-			if err := metrics.ValidateMetricsCommitSynced(reconcilerName, nt.RootRepos[syncName].Hash()); err != nil {
-				return err
-			}
-		}
-
-		for syncNN := range nt.NonRootRepos {
-			reconcilerName := core.NsReconcilerName(syncNN.Namespace, syncNN.Name)
-			if err := metrics.ValidateMetricsCommitSynced(reconcilerName, nt.NonRootRepos[syncNN].Hash()); err != nil {
-				return err
-			}
-		}
-
-		nt.DebugLog(`Found "last_sync_timestamp" metric with commit="<latest>" and status="<any>" for all active reconcilers`)
-		return nil
-	}
-}
-
-// SyncMetricsToLatestCommitSyncedWithSuccess syncs metrics to the latest
-// commit that was applied without errors.
-func SyncMetricsToLatestCommitSyncedWithSuccess(nt *NT) MetricsSyncOption {
-	return func(metrics *testmetrics.ConfigSyncMetrics) error {
-		for syncName := range nt.RootRepos {
-			reconcilerName := core.RootReconcilerName(syncName)
-			if err := metrics.ValidateMetricsCommitSyncedWithSuccess(reconcilerName, nt.RootRepos[syncName].Hash()); err != nil {
-				return err
-			}
-		}
-
-		for syncNN := range nt.NonRootRepos {
-			reconcilerName := core.NsReconcilerName(syncNN.Namespace, syncNN.Name)
-			if err := metrics.ValidateMetricsCommitSyncedWithSuccess(reconcilerName, nt.NonRootRepos[syncNN].Hash()); err != nil {
-				return err
-			}
-		}
-
-		nt.DebugLog(`Found "last_sync_timestamp" metric with commit="<latest>" and status="success" for all active reconcilers`)
-		return nil
-	}
-}
-
-// SyncMetricsToReconcilerSourceError syncs metrics to a reconciler source error
-func SyncMetricsToReconcilerSourceError(nt *NT, reconcilerName string) MetricsSyncOption {
-	ntPtr := nt
-	reconcilerNameCopy := reconcilerName
-	return func(metrics *testmetrics.ConfigSyncMetrics) error {
-		pod := ntPtr.GetDeploymentPod(reconcilerNameCopy, configmanagement.ControllerNamespace)
-		err := metrics.ValidateReconcilerErrors(pod.Name, 1, 0)
-		if err != nil {
-			return err
-		}
-		ntPtr.DebugLog(`Found "reconciler_errors" metric with component="source" and value="1"`)
-		return nil
-	}
-}
-
-// SyncMetricsToReconcilerSyncError syncs metrics to a reconciler sync error
-func SyncMetricsToReconcilerSyncError(nt *NT, reconcilerName string) MetricsSyncOption {
-	ntPtr := nt
-	reconcilerNameCopy := reconcilerName
-	return func(metrics *testmetrics.ConfigSyncMetrics) error {
-		pod := ntPtr.GetDeploymentPod(reconcilerNameCopy, configmanagement.ControllerNamespace)
-		err := metrics.ValidateReconcilerErrors(pod.Name, 0, 1)
-		if err != nil {
-			return err
-		}
-		ntPtr.DebugLog(`Found "reconciler_errors" metric with component="sync" and value="1"`)
-		return nil
-	}
-}
-
-type waitForRepoSyncsOptions struct {
-	timeout            time.Duration
-	syncNamespaceRepos bool
-	rootSha1Fn         Sha1Func
-	repoSha1Fn         Sha1Func
-	syncDirectoryMap   map[types.NamespacedName]string
-}
-
-func newWaitForRepoSyncsOptions(timeout time.Duration, rootFn, repoFn Sha1Func) waitForRepoSyncsOptions {
-	return waitForRepoSyncsOptions{
-		timeout:            timeout,
-		syncNamespaceRepos: true,
-		rootSha1Fn:         rootFn,
-		repoSha1Fn:         repoFn,
-		syncDirectoryMap:   map[types.NamespacedName]string{},
-	}
-}
-
-// WaitForRepoSyncsOption is an optional parameter for WaitForRepoSyncs.
-type WaitForRepoSyncsOption func(*waitForRepoSyncsOptions)
-
-// WithTimeout provides the timeout to WaitForRepoSyncs.
-func WithTimeout(timeout time.Duration) WaitForRepoSyncsOption {
-	return func(options *waitForRepoSyncsOptions) {
-		options.timeout = timeout
-	}
-}
-
-// Sha1Func is the function type that retrieves the commit sha1.
-type Sha1Func func(nt *NT, nn types.NamespacedName) (string, error)
-
-// WithRootSha1Func provides the function to get RootSync commit sha1 to WaitForRepoSyncs.
-func WithRootSha1Func(fn Sha1Func) WaitForRepoSyncsOption {
-	return func(options *waitForRepoSyncsOptions) {
-		options.rootSha1Fn = fn
-	}
-}
-
-// WithRepoSha1Func provides the function to get RepoSync commit sha1 to WaitForRepoSyncs.
-func WithRepoSha1Func(fn Sha1Func) WaitForRepoSyncsOption {
-	return func(options *waitForRepoSyncsOptions) {
-		options.repoSha1Fn = fn
-	}
-}
-
-// RootSyncOnly specifies that only the root-sync repo should be synced.
-func RootSyncOnly() WaitForRepoSyncsOption {
-	return func(options *waitForRepoSyncsOptions) {
-		options.syncNamespaceRepos = false
-	}
-}
-
-// SyncDirPredicatePair is a pair of the sync directory and the predicate.
-type SyncDirPredicatePair struct {
-	Dir       string
-	Predicate func(string) Predicate
-}
-
-// WithSyncDirectoryMap provides a map of RootSync|RepoSync and the corresponding sync directory.
-// The function is used to get the sync directory based on different RootSync|RepoSync name.
-func WithSyncDirectoryMap(syncDirectoryMap map[types.NamespacedName]string) WaitForRepoSyncsOption {
-	return func(options *waitForRepoSyncsOptions) {
-		options.syncDirectoryMap = syncDirectoryMap
-	}
-}
-
-// syncDirectory returns the sync directory if RootSync|RepoSync is in the map.
-// Otherwise, it returns the default sync directory: acme.
-func syncDirectory(syncDirectoryMap map[types.NamespacedName]string, nn types.NamespacedName) string {
-	if syncDir, found := syncDirectoryMap[nn]; found {
-		return syncDir
-	}
-	return AcmeDir
 }
 
 // RemoteRootRepoSha1Fn returns .status.lastSyncedCommit as the latest sha1.
